# back/public/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API ProjetPro'
  description: "API REST pour l'application ProjetPro avec système d'authentification sécurisé.\n\n## Fonctionnalités principales:\n- Authentification avec protection anti-force brute\n- Gestion des tokens CSRF\n- Upload d'images vers Azure Blob Storage\n- Logging automatique des tentatives de connexion\n- Système de blocage par IP et par login\n\n## Sécurité:\n- Protection CSRF obligatoire\n- Tokens d'authentification JWT\n- Blocage automatique après échecs répétés\n\n## Headers requis:\n- `X-CSRF-Token`: Token CSRF pour les requêtes POST\n- `Authorization: Bearer <token>`: Token JWT pour l'authentification\n- `Content-Type: application/json`\n"
  contact:
    name: 'Support API'
    email: support@projetpro.com
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Serveur de développement'
  -
    url: /api
    description: 'API Production'
paths:
...
  /api/images/upload-url:
    post:
      tags:
        - Images
      summary: "Générer une URL d'upload Azure"
      description: 'Génère une URL présignée pour uploader directement une image vers Azure Blob Storage'
      operationId: generateImageUploadUrl
      requestBody:
        description: 'Informations de fichier à uploader'
        required: true
        content:
          application/json:
            schema:
              required:
                - filename
              properties:
                filename:
                  description: 'Nom du fichier avec extension'
                  type: string
                  example: photo.jpg
                expiry:
                  description: 'Durée de validité en secondes'
                  type: integer
                  example: 3600
              type: object
      responses:
        '200':
          description: "URL d'upload générée avec succès"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AzureUploadResponseDTO'
        '400':
          description: 'Paramètres invalides ou extension non autorisée'
  /api/images/verify:
    post:
      tags:
        - Images
      summary: "Vérifier un upload d'image"
      description: "Vérifie qu'un upload a bien eu lieu sur Azure Blob Storage et retourne l'URL publique de l'image."
      operationId: verifyImageUpload
      requestBody:
        description: 'Nom du blob à vérifier'
        required: true
        content:
          application/json:
            schema:
              required:
                - blob_name
              properties:
                blob_name:
                  description: 'Nom du blob/image'
                  type: string
                  example: photo.jpg
              type: object
      responses:
        '200':
          description: 'Image trouvée, URL retournée'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  image_url: { type: string, example: 'https://storageaccount.blob.core.windows.net/container/photo.jpg' }
                  blob_name: { type: string, example: photo.jpg }
                type: object
        '400':
          description: 'Paramètre blob_name manquant'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'blob_name is required' }
                type: object
        '404':
          description: 'Image non trouvée ou upload échoué'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: 'Image not found or upload failed' }
                type: object
  '/api/images/signed-url/{blobName}':
    get:
      tags:
        - Images
      summary: 'Obtenir une URL signée pour une image privée'
      description: 'Génère une URL temporaire signée pour accéder à une image privée sur Azure Blob Storage.'
      operationId: getSignedImageUrl
      parameters:
        -
          name: blobName
          in: path
          description: 'Nom du blob/image'
          required: true
          schema:
            type: string
            example: photo.jpg
        -
          name: expiry
          in: query
          description: "Durée de validité de l'URL en secondes"
          required: false
          schema:
            type: integer
            example: 3600
      responses:
        '200':
          description: 'URL signée générée'
          content:
            application/json:
              schema:
                properties:
                  signed_url: { type: string, example: 'https://...' }
                  expires_at: { type: string, example: '2025-08-11T13:00:00Z' }
                type: object
        '500':
          description: "Erreur lors de la génération de l'URL"
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Failed to generate signed URL' }
                  message: { type: string, example: ... }
                type: object
  '/api/images/{blobName}':
    delete:
      tags:
        - Images
      summary: 'Supprimer une image'
      description: "Supprime une image d'Azure Blob Storage."
      operationId: deleteImage
      parameters:
        -
          name: blobName
          in: path
          description: 'Nom du blob/image à supprimer'
          required: true
          schema:
            type: string
            example: photo.jpg
      responses:
        '200':
          description: 'Image supprimée'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: 'Image deleted successfully' }
                type: object
        '404':
          description: 'Image non trouvée'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: 'Failed to delete image' }
                type: object
        '500':
          description: 'Erreur serveur'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Failed to delete image' }
                  message: { type: string, example: ... }
                type: object
  '/api/images/url/{blobName}':
    get:
      tags:
        - Images
      summary: "Obtenir l'URL publique d'une image"
      description: "Retourne l'URL publique d'une image stockée sur Azure Blob Storage."
      operationId: getPublicImageUrl
      parameters:
        -
          name: blobName
          in: path
          description: 'Nom du blob/image'
          required: true
          schema:
            type: string
            example: photo.jpg
      responses:
        '200':
          description: 'URL publique retournée'
          content:
            application/json:
              schema:
                properties:
                  image_url: { type: string, example: 'https://storageaccount.blob.core.windows.net/container/photo.jpg' }
                  blob_name: { type: string, example: photo.jpg }
                type: object
...
components:
  schemas:
...
    AzureUploadResponseDTO:
      title: "Réponse d'upload Azure"
      description: 'Informations pour uploader directement vers Azure Blob Storage'
      required:
        - upload_url
        - blob_name
        - expires_at
        - headers
        - expired
        - timeUntilExpiry
      properties:
        upload_url:
          description: "URL présignée pour l'upload"
          type: string
          example: 'https://storage.blob.core.windows.net/container/blob?sig=...'
        blob_name:
          description: 'Nom du blob dans Azure'
          type: string
          example: images/uuid-photo.jpg
        expires_at:
          description: "Date d'expiration de l'URL"
          type: string
          format: date-time
          example: '2025-06-29T15:30:00Z'
        headers:
          description: "Headers requis pour l'upload"
          type: object
          example:
            x-ms-blob-type: BlockBlob
            Content-Type: image/jpeg
        expired:
          title: "Vérifie si l'URL d'upload est encore valide."
          type: boolean
        timeUntilExpiry:
          title: 'Retourne le temps restant avant expiration en secondes.'
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Token JWT d'authentification"
      bearerFormat: JWT
      scheme: bearer
    csrfToken:
      type: apiKey
      description: 'Token CSRF pour la protection contre les attaques CSRF'
      name: X-CSRF-Token
      in: header
security:
  -
    bearerAuth: []
  -
    csrfToken: []
...